finished testdriver
(note: originally used wrong file -_-)

Also, parsing code in fraction.cpp was an initial attempt, but buggy
and never finished so I junked it. The replacement code is poor
(compared to repl.cpp) but meets all of the requirements and was
significantly less work to add than abstracting out the parser from
repl atm.